litrpg-notifier/vercel/requirements.txt
aiohttp
requests
python-dotenv
gspread
oauth2client
